# apps/backend/src/routes/cc-genre-artwork.ts
/**
 * @file Serves procedurally generated Creative-Commons genre artworks.
 *
 * The CC browse grid references one of these per tile via `artworkUrl`. The
 * tile uses the shared Spotify-style composite — flat fill in the cover's
 * average colour, genre name baked upper-left, and a rotated cover thumbnail
 * tucked into the lower-right — identical to the commercial tiles. The CC path
 * stays 100% Jamendo: the cover comes from a representative Jamendo album,
 * never Last.fm or Deezer.
 *
 * On a cache hit the JPEG comes straight out of Postgres (keyed `cc:<genreKey>`
 * so CC and commercial artworks never collide on the shared `genre_artworks`
 * table); on a miss the service fetches the Jamendo cover, samples its dominant
 * hue, and synthesises the composite.
 *
 * Marked immutable in `Cache-Control`: the same (genre, algorithm) pair always
 * produces the same bytes, so browsers and CDNs can hold the response
 * indefinitely. Bumping the generator algorithm requires a cache purge or a
 * version-segmented URL (see `CC_ARTWORK_VERSION` in `services/cc/cc-genre.ts`).
 */

import { ROUTE_TEMPLATES } from "@musiccloud/shared";
import type { FastifyInstance } from "fastify";
import { publicErrorResponse } from "../docs/public-response-schema.js";
import { getCcGenreCoverUrl, getCcGenres } from "../services/cc/jamendo/client.js";
import { ensureArtwork, getCachedArtwork } from "../services/genre-artwork/index.js";

const GENRE_KEY_PATTERN = /^[a-z0-9][a-z0-9 &'.\-+]{0,63}$/;

/**
 * Canonicalises a raw genre key from the URL: percent-decoded, lowercased,
 * inner whitespace collapsed, trimmed. Matches the commercial route's
 * normalisation so both caches key on the same shape.
 *
 * @param raw - The raw `:genreKey` path segment.
 * @returns The normalised genre key.
 */
function normaliseGenreKey(raw: string): string {
  return decodeURIComponent(raw).toLowerCase().replace(/\s+/g, " ").trim();
}

/**
 * Title-cases a genre key for the baked-in label when the genre is not in
 * Jamendo's curated station list (fallback only — the station `displayName`
 * is preferred when available).
 *
 * @param genreKey - The normalised genre key.
 * @returns The title-cased display name.
 */
function toDisplayName(genreKey: string): string {
  return genreKey.replace(/\b\w/g, (c) => c.toUpperCase());
}

/**
 * Registers `GET /api/v1/cc/genre-artwork/:genreKey`.
 *
 * Mirrors `routes/genre-artwork.ts` structurally but is Jamendo-fed. Like the
 * commercial route it does NOT call `apiRateLimiter`, so it is exempt from the
 * 10 requests / 60 seconds per-IP quota — the frontend loads CC tiles in
 * parallel from the browse grid. The global 300 requests/minute ceiling still
 * applies.
 *
 * @param app - The Fastify instance to register the route on.
 */
export default async function ccGenreArtworkRoutes(app: FastifyInstance) {
  app.get<{ Params: { genreKey: string } }>(
    ROUTE_TEMPLATES.v1.ccGenreArtwork,
    {
      schema: {
        tags: ["Artwork"],
        summary: "Procedurally generated Creative-Commons genre artwork",
        description:
          "Returns a `512×512` JPEG derived only from Creative Commons source artwork. A stable generated image is served with `Cache-Control: public, max-age=31536000, immutable`. If source artwork is temporarily unavailable, the endpoint still returns a fallback JPEG with `Cache-Control: no-store`, allowing a later request to retry generation. This route is subject only to the global limit of `300` requests in a rolling `60`-second window per client IP.",
        params: {
          type: "object",
          required: ["genreKey"],
          properties: {
            genreKey: {
              type: "string",
              description:
                "Genre name from `CcGenreBrowseResponse.genres[].name`. URL-encode it when inserting it into the path; the endpoint lowercases and normalizes whitespace before validation.",
            },
          },
        },
        response: {
          200: {
            description: "Generated, cached, or transient fallback Creative Commons genre artwork as raw JPEG bytes.",
            headers: {
              "Cache-Control": {
                type: "string",
                description:
                  "`public, max-age=31536000, immutable` for a stable image, or `no-store` for a transient fallback.",
              },
            },
            content: { "image/jpeg": { schema: { type: "string", format: "binary" } } },
          },
          400: publicErrorResponse("The normalized genre key contains invalid characters."),
        },
      },
    },
    async (request, reply) => {
      const genreKey = normaliseGenreKey(request.params.genreKey);
      if (!GENRE_KEY_PATTERN.test(genreKey)) {
        return reply.code(400).send({ error: "INVALID_GENRE_KEY", message: "Genre key contains invalid characters." });
      }

      const cacheKey = `cc:${genreKey}`;
      const cached = await getCachedArtwork(cacheKey);
      if (cached) {
        return reply
          .code(200)
          .header("Content-Type", "image/jpeg")
          .header("Cache-Control", "public, max-age=31536000, immutable")
          .send(cached.jpeg);
      }

      // Prefer the curated station's cleaned label (e.g. "Drum & Bass") over a
      // title-cased key — `getCcGenres` is memoized, so this never refetches
      // `/radios` per tile.
      const displayName =
        (await getCcGenres()).find((g) => g.name === genreKey)?.displayName ?? toDisplayName(genreKey);
      // No cover? `ensureArtwork` tolerates a null cover and renders a
      // flat-colour tile with the name baked in, so the grid never shows a
      // broken tile (no 404). Only refuse on a genuinely nonsense key (above).
      const coverUrl = await getCcGenreCoverUrl(genreKey);
      const { jpeg, isFallback } = await ensureArtwork(cacheKey, coverUrl, displayName);
      return (
        reply
          .code(200)
          .header("Content-Type", "image/jpeg")
          // A transient fallback (cover fetch failed) must NOT be cached
          // immutably or it freezes in the browser; serve it `no-store` so the
          // next view re-requests and retries the cover.
          .header("Cache-Control", isFallback ? "no-store" : "public, max-age=31536000, immutable")
          .send(jpeg)
      );
    },
  );
}