# apps/developer/src/lib/__fixtures__/public-openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Standard public error envelope. Route-specific error schemas add documented fields such as form validation issues or readiness details.",
        "example": {
          "context": {
            "limit": 10,
            "retryAfterSeconds": 42,
            "windowSeconds": 60
          },
          "error": "MC-API-0003",
          "errorId": "7d33e012-685f-4f11-94da-c6bc72918d7b",
          "message": "Too many requests. You can make 10 requests per 60 seconds. Please try again in 42 seconds. (MC-API-0003)"
        },
        "properties": {
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "description": "Optional structured values associated with the error code. For `MC-API-0003`, the object can contain the active limit, window, and retry delay. The key is omitted when no structured values apply.",
            "properties": {
              "limit": {
                "description": "Maximum requests allowed by the rate-limit rule that rejected this request. The key is omitted for non-rate-limit errors.",
                "type": "number"
              },
              "retryAfterSeconds": {
                "description": "Seconds until the client can retry after a `429 Too Many Requests` response. The key is omitted when no retry delay applies.",
                "type": "number"
              },
              "windowSeconds": {
                "description": "Length in seconds of the rate-limit window that rejected this request. The key is omitted for non-rate-limit errors.",
                "type": "number"
              }
            },
            "type": "object"
          },
          "error": {
            "description": "Stable musiccloud error code for programmatic handling, for example `MC-URL-0003`, `MC-API-0003`, or `MC-RES-0001`.",
            "type": "string"
          },
          "errorId": {
            "description": "Unique incident identifier included in the matching backend log. Send it to musiccloud support when reporting the failure.",
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "description": "Safe English failure detail. The final parenthesized value repeats the `error` code.",
            "type": "string"
          }
        },
        "required": ["error", "message", "errorId"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "title": "musiccloud API",
    "version": "2.1.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/cc/genre-artwork/{genreKey}": {
      "get": {
        "description": "Returns a `512×512` JPEG derived only from Creative Commons source artwork. A stable generated image is served with `Cache-Control: public, max-age=31536000, immutable`. If source artwork is temporarily unavailable, the endpoint still returns a fallback JPEG with `Cache-Control: no-store`, allowing a later request to retry generation. This route is subject only to the global limit of `300` requests in a rolling `60`-second window per client IP.",
        "parameters": [
          {
            "description": "Genre name from `CcGenreBrowseResponse.genres[].name`. URL-encode it when inserting it into the path; the endpoint lowercases and normalizes whitespace before validation.",
            "in": "path",
            "name": "genreKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated, cached, or transient fallback Creative Commons genre artwork as raw JPEG bytes.",
            "headers": {
              "Cache-Control": {
                "description": "`public, max-age=31536000, immutable` for a stable image, or `no-store` for a transient fallback.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The normalized genre key contains invalid characters."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Global rate limit exceeded. Retry after the interval provided by the `Retry-After` header.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Use `errorId` from the response when reporting the failure."
          }
        },
        "summary": "Procedurally generated Creative-Commons genre artwork",
        "tags": ["Artwork"]
      }
    },
...
    "/api/v1/genre-artwork/{genreKey}": {
      "get": {
        "description": "Returns a `512×512` JPEG for a normalized genre. A stable generated image is served with `Cache-Control: public, max-age=31536000, immutable`. If source artwork is temporarily unavailable, the endpoint still returns a fallback JPEG with `Cache-Control: no-store`, allowing a later request to retry generation. This route is subject only to the global limit of `300` requests in a rolling `60`-second window per client IP.",
        "parameters": [
          {
            "description": "Genre name from `GenreBrowseResponse.genres[].name`. URL-encode it when inserting it into the path; the endpoint lowercases and normalizes whitespace before validation.",
            "in": "path",
            "name": "genreKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated, cached, or transient fallback genre artwork as raw JPEG bytes.",
            "headers": {
              "Cache-Control": {
                "description": "`public, max-age=31536000, immutable` for a stable image, or `no-store` for a transient fallback.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The normalized genre key contains invalid characters."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Global rate limit exceeded. Retry after the interval provided by the `Retry-After` header.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Use `errorId` from the response when reporting the failure."
          }
        },
        "summary": "Procedurally generated genre artwork",
        "tags": ["Artwork"]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.musiccloud.io"
    }
  ],
...
}